+++ ft_printf/ft_strncmp.c
@@ -6,13 +6,13 @@
 /*   By: yidemir <yidemir@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/24 18:25:19 by yidemir           #+#    #+#             */
-/*   Updated: 2024/11/24 14:20:21 by yidemir          ###   ########.fr       */
+/*   Updated: 2024/11/30 14:04:24 by yidemir          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libftprintf.h"
+#include "ft_printf.h"
 
-int	ft_strncmp(const char *s1, const char *s2, size_t n)
+int	ft_strncmp(const char *s1, const char *s2, int n)
 {
 	unsigned char	c1;
 	unsigned char	c2;
